services/veri-aidpo-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.11.0
transformers==4.46.3
torch==2.5.1
datasets==3.1.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
loguru==0.7.3
redis==5.2.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
